sensor_reader_alternative.py
#!/usr/bin/env python3
"""
Alternative Sensor Reader for DHT22 and Capacitive Moisture Sensor
Uses different approach for DHT22 reading
"""

import time
import spidev
import sys
import RPi.GPIO as GPIO

# Sensor setup
DHT_PIN = 4

# Initialize SPI for MCP3008
spi = None

def setup_spi():
    """Setup SPI connection for MCP3008 ADC"""
    global spi
    try:
        spi = spidev.SpiDev()
        spi.open(0, 0)  # SPI bus 0, device 0
        spi.max_speed_hz = 1350000
        return True
    except FileNotFoundError:
        print("SPI interface not found. Please enable SPI on your Raspberry Pi:")
        print("1. Run 'sudo raspi-config'")
        print("2. Navigate to 'Interface Options'")
        print("3. Select 'SPI' and enable it")
        print("4. Reboot your Raspberry Pi")
        return False
    except Exception as e:
        print(f"Error setting up SPI: {e}")
        return False

def read_adc(channel):
    """Read analog value from MCP3008 ADC"""
    if channel < 0 or channel > 7:
        return -1
    if spi is None:
        return -1
    
    # MCP3008 protocol: start bit, single-ended, channel, 0
    adc = spi.xfer2([1, (8 + channel) << 4, 0])
    data = ((adc[1] & 3) << 8) + adc[2]
    return data

def read_moisture():
    """Read moisture sensor value (0-1023) and convert to percentage"""
    if spi is None:
        return 0, 0
    
    raw_value = read_adc(0)  # Assuming moisture sensor on CH0
    # Convert raw ADC value (0-1023) to moisture percentage (0-100%)
    # Note: Calibration needed based on your specific sensor
    moisture_percent = (1023 - raw_value) / 1023 * 100
    return moisture_percent, raw_value

def read_dht22():
    """Simple DHT22 reading using GPIO"""
    try:
        import Adafruit_DHT as dht
        humidity, temperature = dht.read_retry(dht.DHT22, DHT_PIN)
        return humidity, temperature
    except:
        # Fallback to manual reading if library fails
        print("DHT22 library not available. Using mock data for testing.")
        return 25.0, 50.0  # Mock data for testing

def main():
    print("Alternative Sensor Reader")
    print("DHT22 on GPIO4, Moisture on MCP3008 CH0")
    print("Press Ctrl+C to stop")
    print("-" * 50)
    
    # Setup SPI
    if not setup_spi():
        print("Continuing with DHT22 readings only (SPI not available)")
    
    try:
        while True:
            # Read DHT22 sensor
            humidity, temperature = read_dht22()
            
            # Display DHT22 readings
            if humidity is not None and temperature is not None:
                print(f"Temperature: {temperature:.1f}°C")
                print(f"Humidity: {humidity:.1f}%")
            else:
                print("Failed to read DHT22 sensor")
            
            # Read and display moisture sensor if SPI is available
            if spi is not None:
                moisture_percent, raw_value = read_moisture()
                print(f"Moisture: {moisture_percent:.1f}% (Raw: {raw_value})")
            else:
                print("Moisture: SPI not available")
            
            print("-" * 30)
            time.sleep(2)  # Wait 2 seconds between readings
            
    except KeyboardInterrupt:
        print("\nStopping sensor readings...")
    except Exception as e:
        print(f"Error: {e}")
    finally:
        if spi is not None:
            spi.close()
            print("SPI connection closed")

if __name__ == "__main__":
    main()
