requirements.txt
Adafruit_DHT==1.4.0
spidev==3.5
RPi.GPIO==0.7.1
